Move `test_utils` to `tests/utils` - Also move `ServerRunner` to `tests/utils`
